+++ config/config/src/Config.ts
@@ -22,6 +22,7 @@ export interface Config {
   excludeLinksFromLockfile: boolean
   extraBinPaths: string[]
   extraEnv: Record<string, string>
+  failIfNoMatch: boolean
   filter: string[]
   filterProd: string[]
   rawLocalConfig: Record<string, any>, // eslint-disable-line
@@ -184,6 +185,7 @@ export interface Config {
   gitBranchLockfile?: boolean
   globalDir?: string
   lockfile?: boolean
+  dedupeInjectedDeps?: boolean
 }
 
 export interface ConfigWithDeprecatedSettings extends Config {