+++ store/plugin-commands-store-inspecting/src/index.ts
@@ -1,0 +1,5 @@
+import * as catIndex from './catIndex'
+import * as catFile from './catFile'
+import * as findHash from './findHash'
+
+export { catIndex, catFile, findHash }