+++ config/plugin-commands-config/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/plugin-commands-config
 
+## 2.0.32
+
+### Patch Changes
+
+- Updated dependencies [4e71066dd]
+  - @pnpm/config@20.3.0
+  - @pnpm/cli-utils@2.1.6
+
+## 2.0.31
+
+### Patch Changes
+
+- Updated dependencies [672c559e4]
+  - @pnpm/config@20.2.0
+  - @pnpm/cli-utils@2.1.5
+
 ## 2.0.30
 
 ### Patch Changes