+++ fetching/directory-fetcher/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @pnpm/directory-fetcher
 
+## 7.0.9
+
+### Patch Changes
+
+- Updated dependencies [4d34684f1]
+  - @pnpm/types@9.4.2
+  - @pnpm/fetcher-base@15.0.6
+  - @pnpm/read-project-manifest@5.0.10
+  - @pnpm/resolver-base@11.0.2
+
+## 7.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @pnpm/types@9.4.1
+  - @pnpm/fetcher-base@15.0.5
+  - @pnpm/read-project-manifest@5.0.9
+  - @pnpm/resolver-base@11.0.1
+
 ## 7.0.7
 
 ### Patch Changes