+++ fs/find-packages/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/fs.find-packages
 
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [4d34684f1]
+  - @pnpm/types@9.4.2
+  - @pnpm/read-project-manifest@5.0.10
+
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @pnpm/types@9.4.1
+  - @pnpm/read-project-manifest@5.0.9
+
 ## 2.0.8
 
 ### Patch Changes