+++ reviewing/plugin-commands-licenses/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @pnpm/plugin-commands-licenses
 
+## 3.0.10
+
+### Patch Changes
+
+- Updated dependencies [4e71066dd]
+  - @pnpm/common-cli-options-help@1.1.0
+  - @pnpm/config@20.3.0
+  - @pnpm/cli-utils@2.1.6
+  - @pnpm/license-scanner@2.2.8
+  - @pnpm/lockfile-file@8.1.6
+
+## 3.0.9
+
+### Patch Changes
+
+- Updated dependencies [672c559e4]
+  - @pnpm/config@20.2.0
+  - @pnpm/lockfile-file@8.1.5
+  - @pnpm/license-scanner@2.2.7
+  - @pnpm/cli-utils@2.1.5
+
+## 3.0.8
+
+### Patch Changes
+
+- @pnpm/license-scanner@2.2.6
+
 ## 3.0.7
 
 ### Patch Changes